spliting the shopping cart route
