+++ .gitignore
@@ -1,2 +1,4 @@
 # dependencies
 /node_modules
+
+/nodemon.json
